refactor: Prefer result type for parse attempts
**TL;DR:**
- Add `result` package (not sure if we're `pyproject.toml` or
  `requirements.txt`).
- Update API to have parsers return `Result` types.
- Consolidate unit test cases to make it easier to add test cases with less
  ceremony.
- Add unit tests for other parsers.

The `result` type allows developers to encode in the type signature that a
function may or may not succeed, which is not the case if you're relying on
raising exceptions. It also cleans-up the code at the callsite by removing the
try/catch blocks - allowing consumers to more cleanly handle success/failure
states.

For more information see the official `result` documentation:
https://pypi.org/project/result/